requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain_pinecone
langchain_community
langchain
flask
pypdf
python-dotenv
ipywidgets
accelerate
-e .